feat: preserve OSM node ids by widening route ids to uint64
OSM node ids are 64-bit, so mapping them onto the Route Server's 32-bit
node id lost them and left routes unrequestable by original OSM id. Widen
the node/edge id types to uint64_t across the route graph and the messages
so the OSM loader can keep each node's original id as its route node id.

- types: nodeid/edgeid and RouteRequest ids to uint64_t; GraphToIDMap,
  route_planner, route_tracker and related signatures follow
- nav2_msgs: RouteNode/RouteEdge/EdgeCost, ComputeRoute/ComputeAndTrackRoute
  and DynamicEdges id fields to uint64
- osm loader: assign each node its OSM id as the route node id, key
  graph_to_id_map on the OSM id, and reject negative ids (JOSM temporaries)
  that would collide with the id sentinels
- fix uint64 format strings (PRIu64) in path_converter, route_server, the
  geojson loader and the operation client
- README: rewrite the OSM section in plainer terms, explain the junction and
  shape-node split with the sample graph, and note that OSM ids are preserved
- tests: assert preserved OSM ids and the id round-trip, add negative-id
  rejection, and widen the findRoute/blocked_ids test signatures

+++ nav2_behavior_tree/include/nav2_behavior_tree/plugins/action/compute_and_track_route_action.hpp
@@ -99,8 +99,8 @@ public:
   {
     return providedBasicPorts(
       {
-        BT::InputPort<unsigned int>("start_id", "ID of the start node"),
-        BT::InputPort<unsigned int>("goal_id", "ID of the goal node"),
+        BT::InputPort<uint64_t>("start_id", "ID of the start node"),
+        BT::InputPort<uint64_t>("goal_id", "ID of the goal node"),
         BT::InputPort<geometry_msgs::msg::PoseStamped>(
           "start",
           "Start pose of the path if overriding current robot pose and using poses over IDs"),
@@ -118,13 +118,13 @@ public:
           "error_code_id", "The compute route error code"),
         BT::OutputPort<std::string>(
           "error_msg", "The compute route error msg"),
-        BT::OutputPort<uint16_t>(
+        BT::OutputPort<uint64_t>(
           "last_node_id",
           "ID of the previous node"),
-        BT::OutputPort<uint16_t>(
+        BT::OutputPort<uint64_t>(
           "next_node_id",
           "ID of the next node"),
-        BT::OutputPort<uint16_t>(
+        BT::OutputPort<uint64_t>(
           "current_edge_id",
           "ID of current edge"),
         BT::OutputPort<nav2_msgs::msg::Route>(
